rename: test폴더로 이동
