+++ flower-map-server/src/app.module.ts
@@ -1,6 +1,4 @@
 import { Module } from '@nestjs/common';
-import { AppController } from './app.controller';
-import { AppService } from './app.service';
 import { FeedsModule } from './modules/feeds/feeds.module';
 import { FlowersModule } from './modules/flowers/flowers.module';
 import { LocationsModule } from './modules/locations/locations.module';
@@ -18,7 +16,5 @@ import { ImagesModule } from './modules/images/images.module';
     LocationsModule,
     ImagesModule
   ],
-  controllers: [AppController],
-  providers: [AppService],
 })
 export class AppModule { }
