+++ flower-map-server/test/unit/feeds/feeds.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { FeedsController } from './feeds.controller';
-import { FeedsService } from './feeds.service';
+import { FeedsController } from '../../../src/modules/feeds/feeds.controller';
+import { FeedsService } from '../../../src/modules/feeds/feeds.service';
 
 describe('FeedsController', () => {
   let controller: FeedsController;
