+++ flower-map-server/test/unit/images/images.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { ImagesController } from './images.controller';
-import { ImagesService } from './images.service';
+import { ImagesController } from '../../../src/modules/images/images.controller';
+import { ImagesService } from '../../../src/modules/images/images.service';
 
 describe('ImagesController', () => {
   let controller: ImagesController;
