+++ flower-map-server/test/unit/locations/locations.controller.spec.ts
@@ -1,6 +1,6 @@
 import { Test, TestingModule } from '@nestjs/testing';
-import { LocationsController } from './locations.controller';
-import { LocationsService } from './locations.service';
+import { LocationsController } from '../../../src/modules/locations/locations.controller';
+import { LocationsService } from '../../../src/modules/locations/locations.service';
 
 describe('LocationsController', () => {
   let controller: LocationsController;
